CAMPAIGN 3 + 4: front-end fidelity, VFS comprehension, core engineering
Large verified batch (full build 0/0, 1335 tests green across 10 suites,
clean-room firewall PASS). Continues the doida.exe reverse end-to-end and
rebuilds the Godot front-end 1:1 with the official client. World scene frozen.

CAMPAIGN 3 (workflow / UI / VFS deep reverse -> specs -> client):
- Comprehended 6 clusters (workflow-spine, UI/HUD, vfs-assetio, lua, sound,
  combat, terrain); login path debugger-verified (0x2B sub-opcode, 17-byte
  zero-padded RSA M, secure 1/4, account\tpassword\tPIN contract).
- Promoted to clean specs: opcodes + packets/login + cmsg_char_*, login.md,
  crypto.md; structs/gucomponent+guwindow, ui_hud_layout (5 panels + full
  152-site HUD sweep); formats/pak (RAW verdict), actormotion, sky; lua-config
  (UTF-8), sound, effect-scheduling, terrain-streaming (synchronous).
- Engineered: char-mgmt packet structs + routing, actormotion catalogue,
  sky.box parser, lua-config reader, login credential cipher (Network.Crypto),
  Godot in-game HUD (5 panels + HP/MP gauge + action/status bars + CenteredModal).

CAMPAIGN 4 (front-end 1:1 with the official client):
- Recovered the exact composition of Login / PIN / Server-List / Char-Select
  (the LoginWindow state machine + SelectWindow): per-widget screen + atlas-UV
  rects, the atlas DDS inventory, the Fisher-Yates PIN keypad, the OpeningWindow
  intro ("red ribbon" crawl), the front-end SFX/BGM, the msg.xdb captions.
- Promoted: frontend_scenes.md §11 (pixel-exact rebuild contract), formats/
  msg_xdb.md, specs/intro_sequence.md, sound.md §15, effects.md §A.15.
- Deep VFS comprehension: mount/open/read pipeline, loader-dispatch (no
  magic-sniff; by call-site/extension), the GHTex cache, the linkage chains,
  the bulk-loader; full VFS structure map (43,347 files / 49 ext / 8 chains).
- Engineered: .xeff parser fix (header 8->32B), MsgXdbCatalog (CP949), the
  faithful Godot front-end rebuild (LoginScreen/PinModal/ServerSelect/
  CharacterSelect + OpeningWindow intro + FrontEndAudio + FrontEndEffectPlayer),
  the layer-04 login flow wired to the real 0x2B credential.

Docs/PLAN.md + Docs/ROADMAP.md track the campaigns; Docs/RE/journal.md carries
the provenance. No literal Korean strings or addresses in committed specs
(captions resolve from the VFS msg.xdb at runtime). Pending: IDB annotation +
names.yaml sync, the bgtexture.lst-vs-.txt arbitration, the VFS deep-dive
promotion, the render-fidelity review.

+++ 05.Presentation/MartialHeroes.Client.Godot/Screens/UiAssetLoader.cs
@@ -144,7 +144,7 @@ public sealed class UiAssetLoader : IDisposable
         MsgXdbCatalog? cat = EnsureMsg();
         if (cat is null) return fallback;
 
-        string? s = cat.GetText(id);
+        string? s = cat.GetText((int)id);
         return string.IsNullOrEmpty(s) ? fallback : s;
     }
